% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdi-config.R
\name{rdi_system_config_unpack}
\alias{rdi_system_config_unpack}
\title{Unpack an RDI System Configuration}
\usage{
rdi_system_config_unpack(system_config)
}
\arguments{
\item{system_config}{The \code{system_config} column from the \code{fixed_leader}
item read from \code{\link[=read_rdi]{read_rdi()}}.}
}
\value{
A data frame with one column per unpacked value.
}
\description{
Unpack an RDI System Configuration
}
\examples{
rdi_system_config_unpack(51777)

}
