% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdi-unpack.R
\name{rdi_unpack_system_config}
\alias{rdi_unpack_system_config}
\alias{rdi_unpack_coord_transform}
\title{Unpack an RDI System Configuration}
\usage{
rdi_unpack_system_config(system_config)

rdi_unpack_coord_transform(coord_transform)
}
\arguments{
\item{system_config}{The \code{system_config} column from the \code{fixed_leader}
item read from \code{\link[=read_rdi]{read_rdi()}}.}

\item{coord_transform}{The \code{coord_transform} column from the \code{fixed_leader}
item read from \code{\link[=read_rdi]{read_rdi()}}.}
}
\value{
A data frame with one column per unpacked value.
}
\description{
Unpack an RDI System Configuration
}
\examples{
rdi_unpack_system_config(51777)
rdi_unpack_coord_transform(23)

}
