% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readrdi-package.R
\docType{package}
\name{readrdi-package}
\alias{readrdi}
\alias{readrdi-package}
\title{readrdi: Read 'RDI' Acoustic Doppler Profile Files}
\description{
Functions to inspect 'RDI' (RD Instruments) files
  containing acoustic doppler profile data.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey.dunnington@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
