% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-rdi.R
\name{read_rdi}
\alias{read_rdi}
\alias{rdi_index}
\alias{rdi_detect_data_types}
\title{Read RDI files}
\usage{
read_rdi(
  file,
  index = rdi_index(file),
  types = rdi_detect_data_types(file, index)
)

rdi_index(file, offset = 0, n_max = -1)

rdi_detect_data_types(file, index = rdi_index(file, n_max = 100))
}
\arguments{
\item{file}{A filename from which to read}

\item{index}{An index created using \code{\link[=rdi_index]{rdi_index()}} or a subset thereof.}

\item{types}{The variable IDs to extract from the file as a named
list. Most usefully the output of \code{\link[=rdi_detect_data_types]{rdi_detect_data_types()}}
or a subset thereof.}

\item{offset}{Where to start looking for ensemble start indicators
for \code{\link[=rdi_index]{rdi_index()}} or a vector of ensemble start offsets (probably derived
from \code{\link[=rdi_index]{rdi_index()}}) for \code{\link[=read_rdi]{read_rdi()}}.}

\item{n_max}{Maximum number of ensembles to index}
}
\value{
A \code{list()} with one component for each
}
\description{
Read RDI files
}
\examples{
rdi_file <- system.file("extdata/19101018.rdi", package = "readrdi")
rdi <- read_rdi(rdi_file)

}
\author{
This was written using the explicit and implicit documentation
provided by \code{oce::read.adp.rdi()} written by Daniel Kelley.
}
